requirements.txt
numpy>=1.10.2
scipy>=0.16.1
scikit-learn>=0.17
scikit-image>=0.11.3
matplotlib>=1.5.0
pillow>=3.1.0
